+++ BingoBlockParty/BingoBlockParty.Client/BallGame/Pieces/ClientPeg.cs
@@ -118,7 +118,7 @@ namespace BingoBlockParty.Client.BallGame.Pieces
 
                 var bonusText = "+$" + this.Bonus.ToString();
                 var size = context.MeasureString(bonusText);
-                context.DrawString(bonusText, -size/ 2, -fontSize / 2);
+                context.DrawString(bonusText, (int)-size / 2, -fontSize / 2);
 
 
                 context.Restore();
